+++ Source/UserInputHandler.cpp
@@ -1,7 +1,7 @@
 #include "UserInputHandler.h"
 
 bool UserInputHandler::Buttons[BUTTON_COUNT] = {0};
-int UserInputHandler::buttonPins[BUTTON_COUNT] = {27, 15, 28, 14, 17, 26}; // up, right, down, left, top right, bottom left
+uint8_t UserInputHandler::buttonPins[BUTTON_COUNT] = {27, 15, 28, 14, 17, 26}; // up, right, down, left, top right, bottom left
 
 void UserInputHandler::updateInput() {
     for (int i = 0; i < BUTTON_COUNT; i++) { // NUMBER_OF_BUTTONS defined in UserInput.h
